Decouple plugins asset building from main Webpack config
