+++ package.json
@@ -28,12 +28,14 @@
     "changelog": "./bin/plugin/cli.js changelog",
     "since": "./bin/plugin/cli.js since",
     "readme": "./bin/plugin/cli.js readme",
-    "build": "wp-scripts build",
+    "build": "npm-run-all 'build:!(plugin)'",
+    "build:optimization-detective": "wp-scripts build -c plugins/optimization-detective/webpack.config.js",
     "build-plugins": "npm-run-all 'build:plugin:!(performance-lab)'",
     "build:plugin:performance-lab": "rm -rf build/performance-lab && mkdir -p build/performance-lab && git archive HEAD | tar -x -C build/performance-lab",
     "build:plugin:auto-sizes": "webpack --mode production --env plugin=auto-sizes",
     "build:plugin:dominant-color-images": "webpack --mode production --env plugin=dominant-color-images",
     "build:plugin:embed-optimizer": "webpack --mode production --env plugin=embed-optimizer",
+    "prebuild:plugin:optimization-detective": "npm run build:optimization-detective",
     "build:plugin:optimization-detective": "webpack --mode production --env plugin=optimization-detective",
     "build:plugin:speculation-rules": "webpack --mode production --env plugin=speculation-rules",
     "build:plugin:webp-uploads": "webpack --mode production --env plugin=webp-uploads",
